fix: handle bad reference to OSD
Moves the class definition of GeoTIFFTileSource into the primary exportyed function. This change no longer supports importing the class as is - but can be accessed through OpenSeadragon.GeoTIFFTileSource once the plugin is initialized.
