+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "geotiff-tilesource",
-  "version": "2.0.1",
+  "version": "2.0.2",
   "description": "A plugin tilesource for OpenSeadragon that uses geotiff.js to provide serverless access to view compatible local or remote TIFF files",
   "files": [
     "dist"
